Fix rlmain.cc.
