+++ .github/workflows/lint_cc.yml
@@ -24,5 +24,5 @@ jobs:
     - uses: actions/checkout@v2
     - uses: DoozyX/clang-format-lint-action@v0.12
       with:
-        source: 'win/rl src/nle.c sys/unix/nledl.c include/nle.h include/nledl.h include/nleobs.h'
+        source: 'win/rl src/nle.c include/nle.h include/nleinstance.h include/dloverride.h include/nleobs.h sys/unix/rlmain.cc'
         clangFormatVersion: 12
